// src/enums/table-modal.ts
export enum TableModalTitle {
    Loki = '$loki',
    Assets = 'assets',
    Author = 'author',
    Description = 'description',
    Github = 'github',
    Homepage = 'homepage',
    LastVersion = 'lastversion',
    MainFile = 'mainfile',
    Meta = 'meta',
    Name = 'name',
    Version = 'version',
    Created = 'created',
    Revision = 'revision',
    Files = 'files'
}